FSChunk Threaded (#13)
HttpGw is now able to concurrently retrieve packets of the same request if the file has more than 15MB.
Each thread gets a predetermined list of offsets of the file and puts the respective data chunks on a mapping structure that maps an offset to a chunk of byte data.
Additionally, it has a thread that iterates by the mapping in offset order (1-2-3-...) and waits on the condition of the current offset if it has no data yet. This thread collects those chunks and puts them in order on a single byte array that is sent to the client afterwards.